Publishable trait
